feat: Implemented modal for editing books
Added functionality to the BookListItem component to display a modal for editing book details when the "EDIT" button is clicked. The modal is pre-filled with existing book information and allows users to modify and save the changes. The modal state is controlled using the setShowModal prop passed from the BookListItem component.